More fixed extravagant RAM usage.

+++ utils.py
@@ -62,7 +62,7 @@ def load_all_datas(pathList,labelList):
     y_train = list()
     
     for arr,label in test:
-        x_test.append(arr.astype("float32"))
+        x_test.append(arr)
         y_test.append(label)
     for arr,label in train:
         x_train.append(arr)
